add !important to .nav-link
In the change from v3 to v4,  text in the nav bar became misaligned. I tried a few fixes but this seems to be the simplest and least intrusive.

I tested it across a few of my sites using bootstrap, but I'm not that experienced with css so I'm not sure if I missed any important tests.